Add TestRunner trait to abstract the full test lifecycle
Move ServerBackend from backend.rs to runner.rs, making backend.rs a
pure interface module with no server references. Add TestRunner trait
that drives the test execution lifecycle (iterating test cases, handling
replays, reporting results). ServerTestRunner implements the trait using
the hegel-core event loop. Hegel::run() now delegates to the generic
run_test_case function via TestRunner.
